Code sturcture improvements
