+++ Tests/Resources/V3/basicEmail1.json
@@ -1,0 +1,8 @@
+{
+    "FromEmail":"pilot@mailjet.com",
+    "FromName":"Mailjet Pilot",
+    "Subject":"Your email flight plan!",
+    "Text-part":"Dear passenger, welcome to Mailjet! May the delivery force be with you!",
+    "Html-part":"<h3>Dear passenger, welcome to Mailjet!</h3><br />May the delivery force be with you!",
+    "Recipients":[{"Email":"passenger@mailjet.com"}]
+}
