+++ Tests/Resources/V3/emailheaders.json
@@ -1,0 +1,9 @@
+{
+    "FromEmail":"pilot@mailjet.com",
+    "FromName":"Mailjet Pilot",
+    "Subject":"Your email flight plan!",
+    "Text-part":"Dear passenger, welcome to Mailjet! May the delivery force be with you!",
+    "Html-part":"<h3>Dear passenger, welcome to Mailjet!</h3><br />May the delivery force be with you!",
+    "Recipients":[{"Email":"passenger@mailjet.com"}],
+    "Headers" : {"Reply-To":"copilot@mailjet.com"}
+}
