+++ Tests/Resources/V3/sendingBulk.json
@@ -1,0 +1,20 @@
+{
+    "Messages":[
+        {
+        "FromEmail":"pilot@mailjet.com",
+        "FromName":"Mailjet Pilot",
+        "Recipients":[{"Email":"passenger1@mailjet.com","Name":"passenger 1"}],
+        "Subject":"Your email flight plan!",
+        "Text-part":"Dear passenger 1, welcome to Mailjet! May the delivery force be with you!",
+        "Html-part":"<h3>Dear passenger 1, welcome to Mailjet!</h3><br />May the delivery force be with you!"
+        },
+        {
+        "FromEmail":"pilot@mailjet.com",
+        "FromName":"Mailjet Pilot",
+        "Recipients":[{"Email":"passenger2@mailjet.com","Name":"passenger 2"}],
+        "Subject":"Your email flight plan!",
+        "Text-part":"Dear passenger 2, welcome to Mailjet! May the delivery force be with you!",
+        "Html-part":"<h3>Dear passenger 2, welcome to Mailjet!</h3><br />May the delivery force be with you!"
+        }
+    ]
+}
