+++ Tests/Resources/V3/templateLanguage.json
@@ -1,0 +1,25 @@
+{
+    "FromEmail":"pilot@mailjet.com",
+    "FromName":"Mailjet Pilot",
+    "Subject":"Your email flight plan!",
+    "MJ-TemplateLanguage":true,
+    "Text-part":"Dear passenger, welcome to Mailjet! On this {{var:day:\"monday\"}}, may the delivery force be with you! {{var:personalmessage:\"\"}}",
+    "Html-part":"<h3>Dear passenger, welcome to Mailjet!</h3><br /> On this {{var:day:\"monday\"}}, may the delivery force be with you! {{var:personalmessage:\"\"}}",
+    "Vars":{
+            "day": "Monday"
+    },
+    "Recipients":[
+            {
+                    "Email": "passenger1@mailjet.com",
+                    "Name": "passenger 1",
+                    "Vars": {
+                            "day": "Tuesday",
+                            "personalmessage": "Happy birthday!"
+                    }
+            },
+            {
+                    "Email": "passenger2@mailjet.com",
+                    "Name": "passenger 2"
+            }
+    ]
+}
