+++ Tests/Resources/V31/basicEmail1.json
