+++ Tests/Resources/V31/groupingCampaign.json
